Split out the LockDao fixtures for more reusability
